Allow optional selective access

+++ examples/client/CosemClient.h
@@ -86,13 +86,13 @@ struct Object
 
     void Print()
     {
-        std::cout << "Object " << name << ": " << ln << " Class ID: " << class_id << " Attribute: " << attribute_id << std::endl;
+        std::cout << "Object " << name << ": " << ln << " Class ID: " << class_id << " Attribute: " << (int)attribute_id << std::endl;
     }
 
     std::string name;
     std::string ln;
     std::uint16_t class_id;
-    std::int16_t attribute_id;
+    std::int8_t attribute_id;
 };
